Split off the parts of borrow checking module which should be shared by all extensions.
